Rename(#22): Review 클래스 이름 수정
