+++ src/components/review/thumbnail/TEmotion.tsx
@@ -10,8 +10,8 @@ import Sick from '@/assets/icons/emoji/sick.svg';
 import Smile from '@/assets/icons/emoji/smile.svg';
 import Stareyes from '@/assets/icons/emoji/stareyes.svg';
 import Tired from '@/assets/icons/emoji/tired.svg';
-import Smiley from '@/assets/icons/smiley.svg';
-import { EmotionContent } from '@/objects/Content';
+import Smiley             from '@/assets/icons/smiley.svg';
+import { EmotionContent } from '@/objects/ReviewContent.ts';
 
 type TEmotionProps = {
   thumbnailContent: EmotionContent;
