+++ src/components/review/view/VKpt.tsx
@@ -1,7 +1,7 @@
 import { WhiteContainer } from '@/components/common/Container';
 import { Divider, DividerY } from '@/components/common/Divider';
-import Text from '@/components/common/Text';
-import { KPTContent } from '@/objects/Content';
+import Text           from '@/components/common/Text';
+import { KPTContent } from '@/objects/ReviewContent.ts';
 
 type VKptProps = {
   content: KPTContent;
