+++ src/components/review/view/VText.tsx
@@ -1,6 +1,6 @@
 import { WhiteContainer } from '@/components/common/Container';
-import Text from '@/components/common/Text';
-import { TextContent } from '@/objects/Content';
+import Text            from '@/components/common/Text';
+import { TextContent } from '@/objects/ReviewContent.ts';
 
 type VTextProps = {
   content: TextContent;
