+++ src/components/review/write/R4F.tsx
@@ -2,8 +2,8 @@
 import { WhiteContainer } from '@/components/common/Container';
 import { Divider, DividerY } from '@/components/common/Divider';
 import { InputTextArea } from '@/components/common/InputText';
-import { FourFContent } from '@/objects/Content';
-import useReviewStore from '@/stores/ReviewStore';
+import { FourFContent }  from '@/objects/ReviewContent.ts';
+import useReviewStore    from '@/stores/ReviewStore';
 
 type R4FProps = {
   index: number;
