+++ src/components/review/write/RKpt.tsx
@@ -2,8 +2,8 @@
 import { WhiteContainer } from '@/components/common/Container';
 import { Divider, DividerY } from '@/components/common/Divider';
 import { InputTextArea } from '@/components/common/InputText';
-import { KPTContent } from '@/objects/Content';
-import useReviewStore from '@/stores/ReviewStore';
+import { KPTContent }    from '@/objects/ReviewContent.ts';
+import useReviewStore    from '@/stores/ReviewStore';
 
 type RKptProps = {
   index: number;
