+++ src/pages/Main/Review/ReviewModifyPage.tsx
@@ -3,9 +3,9 @@ import { useParams } from 'react-router-dom';
 
 import getReviewDetail from '@/api/review/getReviewDetail';
 import ViewBlockList from '@/components/review/view/ViewBlockList';
-import WriteBlockList from '@/components/review/write/WriteBlockList';
-import { Content }     from '@/objects/Content';
-import useReviewStore  from '@/stores/ReviewStore';
+import WriteBlockList    from '@/components/review/write/WriteBlockList';
+import { ReviewContent } from '@/objects/ReviewContent.ts';
+import useReviewStore    from '@/stores/ReviewStore';
 
 function ReviewModifyPage() {
   const { scheduleId, reviewId } = useParams();
@@ -17,7 +17,7 @@ function ReviewModifyPage() {
         const response = await getReviewDetail(reviewId);
         const mappedResponse = {
           ...response,
-          contents: response.contents.map((content: Content) => ({
+          contents: response.contents.map((content: ReviewContent) => ({
             ...content,
             blockType: content.blockType as ReviewType,
           })),
